Add SDG English icons in SVG format
So we provide an alternative to the PNG images. SVG icons are smaller,
can be compressed with gzip, and are scalable.

These icons have been downloaded from https://globalgoals.org.
Unfortunately I haven't found SVG files for languages other than
English; on that page they emulate icons in other languages by combining
different SVG files inside the HTML and styling them with CSS so they
look like an icon, but they don't provide fully translated SVG files.

Note we're loading them using `<use>` instead of using separate files
and loading them as images. We're doing so because, when compressed, the
size of the icons is very small (1KB-3KB), and having one request to get
them all (33KB compressed) makes the page load faster. With PNGs we
aren't using a sprite because rendering them in an `<img>` tag would be
harder and the resulting PNG sprite would be too big (about 250KB).

We could also have loaded them inline. However, in this case the icons
would not be cached nor compressed, and they might appear several times
on the same page, making the generated HTML much larger.

Also note we're using the `img` role with an `aria-label` attribute.
This was the only way I found to make my screen reader (Orca) correctly
announce the text when the icon is used inside a link. Using a `<title>`
element inside the SVG didn't work. Regarding this topic, the
documentation on the Mozilla Development Network says [1]:

> If you are using embedded SVG images in your page, it is a good idea
> to set role="img" on the outer <svg> element and give it a label. This
> will cause screen readers to just consider it as a single entity and
> describe it using the label, rather than trying to read out all the
> child nodes

Finally, note that, for reasons I'm not sure about, we need to specify a
viewBox in the SVG element. Otherwise the link will be slightly bigger
than the SVG inside it, which will produce a strange effect when
focusing on the link.

[1] https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Role_Img#svg_and_roleimg